requirements.txt
Jinja2==2.9.6
PyYAML==3.12
oslo.config>=4.6.0
oslo.log>=3.30.0
